switch to spans and remove duplicate tests
